requirements.txt
fastapi
uvicorn
groq
psycopg2-binary
python-dotenv
requests
matplotlib
python-multipart
resend
pillow
pytesseract